+++ Casks/syncthingstatus.rb
@@ -1,6 +1,6 @@
 cask "syncthingstatus" do
   version "1.6.2"
-  sha256 "cd1e15a706605726edf613e9411829e8296d5cf2d0430c6e4362465dc6320f47"
+  sha256 "51641e72c54458147b8ca3c8133b017db3d3b6ad6b24a7366fdbcbe71caf09fe"
 
   url "https://github.com/Xpycode/syncthingStatus/releases/download/v#{version}/syncthingStatus-v#{version}.dmg"
   name "syncthingStatus"